Rename darbelo++'s gitoriusparser to gitoriousparser, fix it up.
The parser doesn't quite work.  I've hacked it up to at least get the
revision hash correct, but that's a hack.  (Currently it just searches
the whole log message for "/commit/[hex stuff]", but it should get the
revision link first and then parse the hash from that instead.)

Three things are still needed for this parser to work correctly:

* Needs to parse the link to the revision from the log content.
* Needs to parse the commit message from the log content.
* Would be nice if it could identify the file(s) changed, either
  directly from the log content (if it's included) or by scraping
  the revision page (like googlecode does).  This isn't a
  requirement, but it's a nice feature.

I've also added a test (using the parrot-plumage repo) to allow
development of the parser.